evm/vm: add EVM interface and use this
