+++ packages/vm/src/types.ts
@@ -1,7 +1,5 @@
 import { Block } from '@ethereumjs/block'
-import { StateAccess } from '@ethereumjs/statemanager'
-import { AccessList, TypedTransaction } from '@ethereumjs/tx'
-import { Address } from 'ethereumjs-util'
+import { TypedTransaction } from '@ethereumjs/tx'
 import { Log } from './evm/types'
 import { AfterBlockEvent } from './runBlock'
 import { AfterTxEvent } from './runTx'
@@ -53,43 +51,4 @@ export type VMEvents = {
   afterBlock: (data: AfterBlockEvent, resolve?: (result: any) => void) => void
   beforeTx: (data: TypedTransaction, resolve?: (result: any) => void) => void
   afterTx: (data: AfterTxEvent, resolve?: (result: any) => void) => void
-}
-
-export interface TransientStorageInterface {
-  get(addr: Address, key: Buffer): Buffer
-  put(addr: Address, key: Buffer, value: Buffer): void
-  commit(): void
-  checkpoint(): void
-  revert(): void
-  toJSON(): { [address: string]: { [key: string]: string } }
-  clear(): void
-}
-
-export interface VmStateAccess extends StateAccess {
-  touchAccount(address: Address): void
-  addWarmedAddress(address: Buffer): void
-  isWarmedAddress(address: Buffer): boolean
-  addWarmedStorage(address: Buffer, slot: Buffer): void
-  isWarmedStorage(address: Buffer, slot: Buffer): boolean
-  clearWarmedAccounts(): void
-  generateAccessList?(addressesRemoved: Address[], addressesOnlyStorage: Address[]): AccessList
-  getOriginalContractStorage(address: Address, key: Buffer): Promise<Buffer>
-  clearOriginalStorageCache(): void
-  cleanupTouchedAccounts(): Promise<void>
-  generateCanonicalGenesis(): Promise<void>
-}
-
-export interface EEIInterface {
-  _transientStorage: TransientStorageInterface
-  state: VmStateAccess
-  getExternalBalance(address: Address): Promise<bigint>
-  getExternalCodeSize(address: bigint): Promise<bigint>
-  getExternalCode(address: bigint): Promise<Buffer>
-  getBlockHash(num: bigint): Promise<bigint>
-  storageStore(address: Address, key: Buffer, value: Buffer): Promise<void>
-  storageLoad(address: Address, key: Buffer, original: boolean): Promise<Buffer>
-  transientStorageStore(address: Address, key: Buffer, value: Buffer): void
-  transientStorageLoad(address: Address, key: Buffer): Buffer
-  isAccountEmpty(address: Address): Promise<boolean>
-  accountExists(address: Address): Promise<boolean>
 }
